Converting to int32

+++ clabtoolkit/parcellationtools.py
@@ -613,7 +613,7 @@ class Parcellation:
         
         # Correcting if new_codes is an integer
         if isinstance(new_codes, int):
-            new_codes = [new_codes]
+            new_codes = [np.int32(new_codes)]
 
         # Detect thecodes2group is a list of list
         if isinstance(codes2rep, list):
